Fixed Realm onThread crashes

+++ app/src/main/java/rx/music/ui/auth/AuthController.kt
@@ -66,7 +66,7 @@ class AuthController : MoxyController(), AuthView {
         loginButton.onClick {
             authPresenter.login(usernameEditText.text.toString(),
                     passwordEditText.text.toString(),
-                    code = validationEditText.text.toString().toInt())
+                    code = validationEditText.text.toString())
         }
     }
